% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_api.R
\name{datadoi}
\alias{datadoi}
\title{Function to extract external ids from lens.org JSON as a single column (including MAG and doi), along with Environmental
Evidence URL for the html full text where provided}
\usage{
datadoi(input)
}
\description{
Function to extract external ids from lens.org JSON as a single column (including MAG and doi), along with Environmental
Evidence URL for the html full text where provided
}
