% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eej_records}
\alias{eej_records}
\title{DATASET_TITLE}
\format{
A data frame with 222 rows and 8 variables:
\describe{
\item{\code{record}}{double COLUMN_DESCRIPTION}
\item{\code{title}}{character COLUMN_DESCRIPTION}
\item{\code{url}}{character COLUMN_DESCRIPTION}
\item{\code{doi}}{character COLUMN_DESCRIPTION}
\item{\code{summary}}{character COLUMN_DESCRIPTION}
\item{\code{authors}}{character COLUMN_DESCRIPTION}
\item{\code{abstract}}{character COLUMN_DESCRIPTION}
\item{\code{type}}{character COLUMN_DESCRIPTION}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
eej_records
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
