% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_texts.R
\name{ensure_library}
\alias{ensure_library}
\title{Choose.directory function for selection of all files in a folder
Thanks to 'dww' for the code: https://stackoverflow.com/questions/48218491/os-independent-way-to-select-directory-interactively-in-r}
\usage{
ensure_library(lib.name)
}
\description{
Choose.directory function for selection of all files in a folder
Thanks to 'dww' for the code: https://stackoverflow.com/questions/48218491/os-independent-way-to-select-directory-interactively-in-r
}
