% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections.R
\name{extract_acknowledgements}
\alias{extract_acknowledgements}
\title{Function to extract Acknowledgements section text}
\usage{
extract_acknowledgements(text)
}
\description{
Function to extract Acknowledgements section text
}
