% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections.R
\name{extract_articletype}
\alias{extract_articletype}
\title{Discern article type from publisher field code}
\usage{
extract_articletype(text)
}
\description{
Discern article type from publisher field code
}
