% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections.R
\name{extract_background}
\alias{extract_background}
\title{Function to extract code for pre-specified sections}
\usage{
extract_background(text, from = "Background", to = "Objective")
}
\description{
Function to extract code for pre-specified sections
}
