% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections.R
\name{extract_contributions}
\alias{extract_contributions}
\title{Function to extract Author's contributions section text}
\usage{
extract_contributions(text)
}
\description{
Function to extract Author's contributions section text
}
