% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections.R
\name{extract_dataavail}
\alias{extract_dataavail}
\title{Function to extract data availability section text}
\usage{
extract_dataavail(text)
}
\description{
Function to extract data availability section text
}
