% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections.R
\name{extract_emails}
\alias{extract_emails}
\title{This file contains a suite of functions to scrape information from web-based HTML and PDF files, and has been developed to function
with the journal Environmental Evidence, published by BioMed Central (https://environmentalevidencejournal.biomedcentral.com/).
Save and scrape full texts from the journal Environmental Evidence.}
\usage{
extract_emails(text)
}
\arguments{
\item{text}{Text extracted from HTML files downloaded from EEJ (see import_text)}
}
\value{
lists of paragraphs from extracted text
Extract email addresses from PDFs
}
\description{
Scrapes the content for html code and full text.
}
