% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections.R
\name{extract_nextpara}
\alias{extract_nextpara}
\title{Function to extract a single paragraph below a given subtitle}
\usage{
extract_nextpara(text, subtitle)
}
\description{
Function to extract a single paragraph below a given subtitle
}
