% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections.R
\name{extract_subsections}
\alias{extract_subsections}
\title{Function to extract code for user-specified sections (based on 'from' and 'to')}
\usage{
extract_subsections(text, from, to)
}
\description{
Function to extract code for user-specified sections (based on 'from' and 'to')
}
