% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections.R
\name{extract_thispara}
\alias{extract_thispara}
\title{Function to extract single para at point of reference}
\usage{
extract_thispara(subtitle, text)
}
\description{
Function to extract single para at point of reference
}
