% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections_many.R
\name{extractall_acknowledgements}
\alias{extractall_acknowledgements}
\title{Function to extract Acknowledgements section texts}
\usage{
extractall_acknowledgements(text)
}
\description{
Function to extract Acknowledgements section texts
}
