% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections_many.R
\name{extractall_articletypes}
\alias{extractall_articletypes}
\title{Discern article type from publisher field code}
\usage{
extractall_articletypes(extract_articletype, text)
}
\description{
Discern article type from publisher field code
}
