% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections_many.R
\name{extractall_backgrounds}
\alias{extractall_backgrounds}
\title{Function to extract code for pre-specified sections from across multiple files}
\usage{
extractall_backgrounds(extract_background, text)
}
\description{
Function to extract code for pre-specified sections from across multiple files
}
