% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections_many.R
\name{extractall_coi}
\alias{extractall_coi}
\title{Function to extract Competing interests section text}
\usage{
extractall_coi(text)
}
\description{
Function to extract Competing interests section text
}
