% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections_many.R
\name{extractall_contributions}
\alias{extractall_contributions}
\title{Function to extract Author's contributions section text}
\usage{
extractall_contributions(text)
}
\description{
Function to extract Author's contributions section text
}
