% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections_many.R
\name{extractall_dataavail}
\alias{extractall_dataavail}
\title{Function to extract data availability section text}
\usage{
extractall_dataavail(text)
}
\description{
Function to extract data availability section text
}
