% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections_many.R
\name{extractall_emails}
\alias{extractall_emails}
\title{This file contains a suite of functions to scrape information from web-based HTML and PDF files, and has been developed to function
with the journal Environmental Evidence, published by BioMed Central (https://environmentalevidencejournal.biomedcentral.com/).
Functions to download and scrape content locally from Environmental Evidence articles based on a doi/publisher url
Save and scrape full texts from the journal Environmental Evidence.}
\usage{
extractall_emails(extract_emails, text)
}
\arguments{
\item{doi}{A single or list of digital object identifier(s) (DOI).}
}
\value{
Locally saved HTML and PDF files, along with scraped text and code from HTMLs and text from PDFs.
Extract email addresses from PDFs
}
\description{
Scrapes the content for html code and full text.
}
