% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections_many.R
\name{extractall_keywords}
\alias{extractall_keywords}
\title{Function to extract Keywords section text}
\usage{
extractall_keywords(text, number = 20)
}
\description{
Function to extract Keywords section text
}
