% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections_many.R
\name{extractall_nextpara}
\alias{extractall_nextpara}
\title{Function to extract single paragraphs below given subtitles}
\usage{
extractall_nextpara(text, subtitle)
}
\description{
Function to extract single paragraphs below given subtitles
}
