% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections_many.R
\name{extractall_subsections}
\alias{extractall_subsections}
\title{Function to extract code for user-specified sections (based on 'from' and 'to') from across multiple files}
\usage{
extractall_subsections(extract_subsections, text, from, to)
}
\description{
Function to extract code for user-specified sections (based on 'from' and 'to') from across multiple files
}
