% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_sections_many.R
\name{extractall_thispara}
\alias{extractall_thispara}
\title{Function to extract single para at point of reference}
\usage{
extractall_thispara(text, subtitle)
}
\description{
Function to extract single para at point of reference
}
