% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_api.R
\name{getScholarlyData}
\alias{getScholarlyData}
\title{Functions to update identification, scraping and subectioning of EEJ systematic reviews and maps}
\usage{
getScholarlyData(token, query)
}
\arguments{
\item{token}{A string provided on approval of access to the lens.org API.}
}
\value{
A dataframe containing the lens.org results, including parsed dois (or other external identifiers) and publisher URLs.
}
\description{
Searches for the source name "Environmental Evidence" are performed using the free API for lens.org.
}
\details{
These functions work to perform a journal name search of the lens.org scholarly API, extract dois and publisher urls for each record,
scrape the html full text content, categorise the results as systematic reviews or systematic maps, and subsection reviews and maps
for full text analysis.

Search lens.org for articles published in the journal Environmental Evidence.
}
