% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfurl.R
\name{pdflink}
\alias{pdflink}
\title{Convert dois to pdflinks}
\usage{
pdflink(doi)
}
\arguments{
\item{doi}{A list of dois}
}
\value{
A list of links for pdfs based on the doi provided
}
\description{
Convert dois to pdflinks
}
