% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfurl.R, R/save_and_scrape.R
\name{pdfurl}
\alias{pdfurl}
\title{Generate URL for PDF location based on doi for EEJ}
\usage{
pdfurl(doi)

pdfurl(doi)
}
\arguments{
\item{doi}{A single or list of digital object identifier(s) (DOI).}
}
\value{
Locally saved HTML and PDF files, along with scraped text and code from HTMLs and text from PDFs.
Generate urls for PDFs based on list of dois
}
\description{
Saves HTML and/or PDF files based on a doi or URL, and scrapes the content for html code and full text.
}
\details{
These functions download single PDF and HTML file for full texts from Environmental Evidence, as well as scraping text
out of an HTML file (including and excluding code) and PDF file (text only), saving the scraped text as .txt files.
The functions have only been verified for articles in Environmental Evidence.

Save and scrape full texts from the journal Environmental Evidence.
}
