% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_scrape_many.R
\name{pdfurls}
\alias{pdfurls}
\title{Functions to download and scrape content locally from Environmental Evidence articles based on a doi/publisher url}
\usage{
pdfurls(doi)
}
\arguments{
\item{doi}{A single or list of digital object identifier(s) (DOI).}
}
\value{
Locally saved HTML and PDF files, along with scraped text and code from HTMLs and text from PDFs.
convert dois to pdflinks
}
\description{
Saves HTML and/or PDF files based on a doi or URL, and scrapes the content for html code and full text.
}
\details{
These functions apply single input functions to a list of dois (see scrape_and_save.R).

Save and scrape full texts from the journal Environmental Evidence.
}
