% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_texts.R
\name{read_texts}
\alias{read_texts}
\title{Running list of dependencies:
rstudioapi - directory selection for read_txt
Function to read in all text files containing full texts}
\usage{
read_texts(arg1 = NULL)
}
\description{
Running list of dependencies:
rstudioapi - directory selection for read_txt
Function to read in all text files containing full texts
}
