% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_text.R
\name{read_txt}
\alias{read_txt}
\title{Running list of dependencies:
rstudioapi - directory selection for read_txt
Function to read in a single text file containing full text scraped from html}
\usage{
read_txt()
}
\description{
Running list of dependencies:
rstudioapi - directory selection for read_txt
Function to read in a single text file containing full text scraped from html
}
