% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_text.R, R/import_texts.R
\name{readin_html}
\alias{readin_html}
\title{Function to read in an html file as code}
\usage{
readin_html()

readin_html()
}
\description{
Function to read in an html file as code

Function to read in all html files in the working directory as code
}
