% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_scrape_many.R
\name{save_pdfs}
\alias{save_pdfs}
\title{Download PDFs to folder (working directory) with doi as the filename (substituting '..' for '/')}
\usage{
save_pdfs(doi)
}
\description{
Download PDFs to folder (working directory) with doi as the filename (substituting '..' for '/')
}
