% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_scrape.R
\name{pdflink}
\alias{pdflink}
\title{Function to download and scrape content locally from Environmental Evidence articles based on a doi/publisher url}
\usage{
pdflink(doi)
}
\arguments{
\item{doi}{A single or list of digital object identifier(s) (DOI).}
}
\value{
Locally saved HTML and PDF files, along with scraped text and code from HTMLs and text from PDFs.
Generate url for PDF
}
\description{
Saves HTML and/or PDF files based on a doi or URL, and scrapes the content for html code and full text.
}
\details{
These functions download PDF and HTML files for full texts from Environmental Evidence, as well as scraping text out of HTML files (including and excluding code)
and PDF files (text only), saving the scraped text as .txt files. The functions only work for Environmental Evidence.

Save and scrape full texts from the journal Environmental Evidence.
}
