% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfurl.R
\name{pdfurl}
\alias{pdfurl}
\title{Convert dois to pdflinks}
\usage{
pdfurl(pdflink, doi)
}
\arguments{
\item{pdflink}{A list of pdflinks from the pdflinks function}

\item{doi}{A list of dois}
}
\value{
A list of links for pdfs based on the doi provided
}
\description{
Convert dois to pdflinks
}
