% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_scrape.R
\name{save_pdf}
\alias{save_pdf}
\title{Download PDFs to folder (working directory) with doi as the filename (substituting '..' for '/')}
\usage{
save_pdf(pdfurl = NULL, doi = NULL)
}
\description{
Download PDFs to folder (working directory) with doi as the filename (substituting '..' for '/')
}
