% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_and_scrape.R
\name{save_pdf}
\alias{save_pdf}
\title{Download PDF to folder (working directory) with doi as the filename (substituting '..' for '/')}
\usage{
save_pdf(doi)
}
\description{
Download PDF to folder (working directory) with doi as the filename (substituting '..' for '/')
}
